sessions: respect reduced motion in chat list spinner
Use the static loading codicon for in-progress sessions when reduced motion is enabled, and rerender the sessions list when that preference changes.

Extract the pure sessions list helpers into a CSS-free module so the list helper tests can cover the reduced-motion icon behavior under the Node test harness.
