+++ src/vs/sessions/contrib/sessions/browser/views/sessionsList.ts
@@ -16,7 +16,7 @@ import { createMatches, FuzzyScore, IMatch } from '../../../../../base/common/fi
 import { Disposable, DisposableStore, MutableDisposable, toDisposable } from '../../../../../base/common/lifecycle.js';
 import { MarkdownString } from '../../../../../base/common/htmlContent.js';
 import { IReader, autorun } from '../../../../../base/common/observable.js';
-import { ThemeIcon, themeColorFromId } from '../../../../../base/common/themables.js';
+import { ThemeIcon } from '../../../../../base/common/themables.js';
 import { URI } from '../../../../../base/common/uri.js';
 import { fromNow } from '../../../../../base/common/date.js';
 import { localize } from '../../../../../nls.js';
@@ -40,10 +40,14 @@ import { Separator } from '../../../../../base/common/actions.js';
 import { IHoverService } from '../../../../../platform/hover/browser/hover.js';
 import { HoverStyle } from '../../../../../base/browser/ui/hover/hover.js';
 import { HoverPosition } from '../../../../../base/browser/ui/hover/hoverWidget.js';
+import { IAccessibilityService } from '../../../../../platform/accessibility/common/accessibility.js';
 import { ISessionsManagementService } from '../../../../services/sessions/common/sessionsManagement.js';
 import { ISessionsListModelService } from './sessionsListModelService.js';
+import { getSessionStatusIcon, groupSessionsForList, ISessionSection, SessionsGrouping, SessionsSorting } from './sessionsListHelpers.js';
 
 const $ = DOM.$;
+
+export { getSessionStatusIcon, groupByDate, groupByWorkspace, groupSessionsForList, ISessionSection, SessionsGrouping, SessionsSorting, sortSessions } from './sessionsListHelpers.js';
 
 export const SessionItemToolbarMenuId = new MenuId('SessionItemToolbar');
 export const SessionItemContextMenuId = new MenuId('SessionItemContextMenu');
@@ -54,22 +58,6 @@ export const IsSessionReadContext = new RawContextKey<boolean>('sessionItem.isRe
 export const SessionSectionTypeContext = new RawContextKey<string>('sessionSection.type', '');
 
 //#region Types
-
-export enum SessionsGrouping {
-	Workspace = 'workspace',
-	Date = 'date',
-}
-
-export enum SessionsSorting {
-	Created = 'created',
-	Updated = 'updated',
-}
-
-export interface ISessionSection {
-	readonly id: string;
-	readonly label: string;
-	readonly sessions: ISession[];
-}
 
 export interface ISessionShowMore {
 	readonly showMore: true;
@@ -172,6 +160,7 @@ class SessionItemRenderer implements ITreeRenderer<SessionListItem, FuzzyScore, 
 		private readonly contextKeyService: IContextKeyService,
 		private readonly markdownRendererService: IMarkdownRendererService,
 		private readonly hoverService: IHoverService,
+		private readonly accessibilityService: IAccessibilityService,
 	) { }
 
 	renderTemplate(container: HTMLElement): ISessionItemTemplate {
@@ -456,20 +445,7 @@ class SessionItemRenderer implements ITreeRenderer<SessionListItem, FuzzyScore, 
 	}
 
 	private getStatusIcon(status: SessionStatus, isRead: boolean, isArchived: boolean, pullRequestIcon?: ThemeIcon): ThemeIcon {
-		switch (status) {
-			case SessionStatus.InProgress: return { ...ThemeIcon.modify(Codicon.loading, 'spin'), color: themeColorFromId('textLink.foreground') };
-			case SessionStatus.NeedsInput: return { ...Codicon.circleFilled, color: themeColorFromId('list.warningForeground') };
-			case SessionStatus.Error: return { ...Codicon.error, color: themeColorFromId('errorForeground') };
-			default:
-				if (pullRequestIcon) {
-					return pullRequestIcon;
-				}
-
-				if (!isRead && !isArchived) {
-					return { ...Codicon.circleFilled, color: themeColorFromId('textLink.foreground') };
-				}
-				return { ...Codicon.circleSmallFilled, color: themeColorFromId('agentSessionReadIndicator.foreground') };
-		}
+		return getSessionStatusIcon(status, isRead, isArchived, this.accessibilityService.isMotionReduced(), pullRequestIcon);
 	}
 
 	private getWorkspaceBadgeLabel(workspace: ISessionWorkspace): string | undefined {
@@ -694,6 +670,7 @@ export class SessionsList extends Disposable implements ISessionsList {
 		@IContextMenuService private readonly contextMenuService: IContextMenuService,
 		@IMenuService private readonly menuService: IMenuService,
 		@IKeybindingService private readonly keybindingService: IKeybindingService,
+		@IAccessibilityService private readonly accessibilityService: IAccessibilityService,
 	) {
 		super();
 
@@ -720,6 +697,7 @@ export class SessionsList extends Disposable implements ISessionsList {
 			contextKeyService,
 			markdownRendererService,
 			hoverService,
+			this.accessibilityService,
 		);
 
 		const showMoreRenderer = new SessionShowMoreRenderer();
@@ -840,6 +818,8 @@ export class SessionsList extends Disposable implements ISessionsList {
 				this.update();
 			}
 		}));
+
+		this._register(this.accessibilityService.onDidChangeReducedMotion(() => this.tree.rerender()));
 
 		this.refresh();
 	}
@@ -1224,132 +1204,6 @@ function getFirstApprovalAcrossChats(approvalModel: AgentSessionApprovalModel, s
 		}
 	}
 	return oldest;
-}
-
-//#endregion
-
-//#region Sorting & Grouping Helpers
-
-export function sortSessions(sessions: ISession[], sorting: SessionsSorting): ISession[] {
-	return [...sessions].sort((a, b) => {
-		if (sorting === SessionsSorting.Updated) {
-			return b.updatedAt.get().getTime() - a.updatedAt.get().getTime();
-		}
-		return b.createdAt.getTime() - a.createdAt.getTime();
-	});
-}
-
-export function groupSessionsForList(
-	sessions: ISession[],
-	grouping: SessionsGrouping,
-	sorting: SessionsSorting,
-	isSessionPinned: (session: ISession) => boolean,
-): ISessionSection[] {
-	const sorted = sortSessions(sessions, sorting);
-
-	// Archived always wins over pinned so done sessions stay grouped together.
-	const pinned: ISession[] = [];
-	const archived: ISession[] = [];
-	const regular: ISession[] = [];
-	for (const session of sorted) {
-		if (session.isArchived.get()) {
-			archived.push(session);
-		} else if (isSessionPinned(session)) {
-			pinned.push(session);
-		} else {
-			regular.push(session);
-		}
-	}
-
-	const sections: ISessionSection[] = [];
-	if (pinned.length > 0) {
-		sections.push({ id: 'pinned', label: localize('pinned', "Pinned"), sessions: pinned });
-	}
-
-	sections.push(...(grouping === SessionsGrouping.Workspace
-		? groupByWorkspace(regular)
-		: groupByDate(regular, sorting)));
-
-	if (archived.length > 0) {
-		sections.push({ id: 'archived', label: localize('archived', "Done"), sessions: archived });
-	}
-
-	return sections;
-}
-
-export function groupByWorkspace(sessions: ISession[]): ISessionSection[] {
-	const groups = new Map<string, ISession[]>();
-	for (const session of sessions) {
-		const workspace = session.workspace.get();
-		const label = workspace?.label || localize('unknown', "Unknown");
-		let group = groups.get(label);
-		if (!group) {
-			group = [];
-			groups.set(label, group);
-		}
-		group.push(session);
-	}
-
-	const unknownWorkspaceLabel = localize('unknown', "Unknown");
-	const order = [...groups.keys()]
-		.filter(k => k !== unknownWorkspaceLabel)
-		.sort((a, b) => a.localeCompare(b));
-
-	const result: ISessionSection[] = order.map(label => ({
-		id: `workspace:${label}`,
-		label,
-		sessions: groups.get(label)!,
-	}));
-
-	// "Unknown Workspace" always at the bottom
-	const unknownWorkspace = groups.get(unknownWorkspaceLabel);
-	if (unknownWorkspace) {
-		result.push({ id: `workspace:${unknownWorkspaceLabel}`, label: unknownWorkspaceLabel, sessions: unknownWorkspace });
-	}
-
-	return result;
-}
-
-export function groupByDate(sessions: ISession[], sorting: SessionsSorting): ISessionSection[] {
-	const now = new Date();
-	const startOfToday = new Date(now.getFullYear(), now.getMonth(), now.getDate()).getTime();
-	const startOfYesterday = startOfToday - 86_400_000;
-	const startOfWeek = startOfToday - 7 * 86_400_000;
-
-	const today: ISession[] = [];
-	const yesterday: ISession[] = [];
-	const week: ISession[] = [];
-	const older: ISession[] = [];
-
-	for (const session of sessions) {
-		const time = sorting === SessionsSorting.Updated
-			? session.updatedAt.get().getTime()
-			: session.createdAt.getTime();
-
-		if (time >= startOfToday) {
-			today.push(session);
-		} else if (time >= startOfYesterday) {
-			yesterday.push(session);
-		} else if (time >= startOfWeek) {
-			week.push(session);
-		} else {
-			older.push(session);
-		}
-	}
-
-	const sections: ISessionSection[] = [];
-	const addGroup = (id: string, label: string, groupSessions: ISession[]) => {
-		if (groupSessions.length > 0) {
-			sections.push({ id, label, sessions: groupSessions });
-		}
-	};
-
-	addGroup('today', localize('today', "Today"), today);
-	addGroup('yesterday', localize('yesterday', "Yesterday"), yesterday);
-	addGroup('thisWeek', localize('lastSevenDays', "Last 7 Days"), week);
-	addGroup('older', localize('older', "Older"), older);
-
-	return sections;
 }
 
 //#endregion
